requirements.txt
fastmcp>=1.0.0
httpx>=0.24.0 